W8D1 Search Functionality

+++ musify/src/main/java/com/example/musify/dto/UserRowMapper.java
@@ -11,7 +11,7 @@ import java.sql.SQLException;
 public class UserRowMapper implements RowMapper<User> {
     public User mapRow(ResultSet rs, int rowNum) throws SQLException {
         User user;
-        user = new User(rs.getInt("id"), rs.getString("first_name"),rs.getString("last_name"),rs.getString("email"),rs.getString("password"),
+        user = new User(rs.getLong("id"), rs.getString("first_name"),rs.getString("last_name"),rs.getString("email"),rs.getString("password"),
                 rs.getString("country_of_origin"),rs.getInt("role"), rs.getInt("active"));
         return user;
     }
